[MTDSA-316] [CK] Make allowances.capitalAllowance and adjustments.lossBroughtForward as optional
