+++ test/uk/gov/hmrc/selfassessmentapi/domain/furnishedholidaylettings/FurnishedHolidayLettingsSpec.scala
@@ -27,15 +27,15 @@ class FurnishedHolidayLettingsSpec extends JsonSpec {
       roundTripJson(FurnishedHolidayLettings(None, "Cosa del Sol apartment", PropertyLocationType.UK, None, None))
 
       roundTripJson(FurnishedHolidayLettings(None, "Cosa del Sol apartment", PropertyLocationType.UK,
-        Some(Allowances(BigDecimal(1000.00))),
-        Some(Adjustments(BigDecimal(500.00)))))
+        Some(Allowances(Some(BigDecimal(1000.00)))),
+        Some(Adjustments(Some(BigDecimal(500.00))))))
     }
 
     "reject name with more than 100 characters" in {
       val fhl = FurnishedHolidayLettings(None, "Abcd" * 100, PropertyLocationType.UK,
-        Some(Allowances(BigDecimal(1000.00))),
-        Some(Adjustments(BigDecimal(500.00))))
-      assertValidationError[FurnishedHolidayLettings](
+        Some(Allowances(Some(BigDecimal(1000.00)))),
+        Some(Adjustments(Some(BigDecimal(500.00)))))
+        assertValidationError[FurnishedHolidayLettings](
         fhl,
         Map(("/name", MAX_FIELD_LENGTH_EXCEEDED) -> "field length exceeded the max 100 chars"),
         "Expected invalid furnished-holiday-lettings")
@@ -45,9 +45,9 @@ class FurnishedHolidayLettingsSpec extends JsonSpec {
     "reject capitalAllowance with negative amounts" in {
       Seq(BigDecimal(-1213.00), BigDecimal(-2243434.00)).foreach { amount =>
         val fhl = FurnishedHolidayLettings(None, "Cosa del Sol apartment", PropertyLocationType.UK,
-          Some(Allowances(amount)),
-          Some(Adjustments(BigDecimal(500.00))))
-        assertValidationError[FurnishedHolidayLettings](
+          Some(Allowances(Some(amount))),
+          Some(Adjustments(Some(BigDecimal(500.00)))))
+          assertValidationError[FurnishedHolidayLettings](
           fhl,
           Map(("/allowances/capitalAllowance", INVALID_MONETARY_AMOUNT) -> "capitalAllowance should be non-negative number up to 2 decimal values"),
           "Expected invalid furnished-holiday-lettings")
@@ -57,9 +57,9 @@ class FurnishedHolidayLettingsSpec extends JsonSpec {
     "reject lossBroughtForward with negative amounts" in {
       Seq(BigDecimal(-1213.00), BigDecimal(-2243434.00)).foreach { amount =>
         val fhl = FurnishedHolidayLettings(None, "Cosa del Sol apartment", PropertyLocationType.UK,
-          Some(Allowances(BigDecimal(500.00))),
-          Some(Adjustments(amount)))
-        assertValidationError[FurnishedHolidayLettings](
+          Some(Allowances(Some(BigDecimal(500.00)))),
+          Some(Adjustments(Some(amount))))
+          assertValidationError[FurnishedHolidayLettings](
           fhl,
           Map(("/adjustments/lossBroughtForward", INVALID_MONETARY_AMOUNT) -> "lossBroughtForward should be non-negative number up to 2 decimal values"),
           "Expected invalid furnished-holiday-lettings")
